refactor: avoid cloning server instance by using Arc<Server>
Currently when creating a server instance and spawing a Tokio task
the server instance is cloned (server derive-implements "Clone" trait).

Instead the same instance is used and an Arc<Server> is handled to the
concurrent Tokio task.

This reduces repetition and memory consumption and keeps the Server
instance unique and consistent through threads.