fix: implementation of file size formatting (#379)
This PR removes `src/utils/fmt.rs`, since it only had the `format_bytes`
function, which can also be used from another library.

The implementation of `format_bytes` is also incorrect. `1024 bytes` is
not equal to `1KB`, as the test indicates. `1024 bytes` is equal to
`1KiB`, and means `Kibibyte`, where `bi` stands for binary. Why not just
use a library for that? That would eliminate such bugs.

The PR changes the file size type from `String` to `u64` and adds a
`_bytes` postfix to the property name.
```diff
- pub(crate) size: String,
+ pub(crate) size_bytes: u64,
```
I would love to store the file size as the `Size` struct provided by the
`size` crate and remove the postfix, but `Size` doesn't implement the
`Serialize` trait, and we'd have to wait for that before we can do that.

---------

Co-authored-by: Esteban Borai <estebanborai@gmail.com>