fix(config): `--spa` implies `--index` option (#422)
If #420 there was still an issue that `--spa would work very weirdly
without `--index`; single page applications are never implemented
without also using `index.html` files. This PR fixes this by
automatically setting `cli.index` to `true` if `cli.spa` is set to true.

I could throw an error if the `--index` flag is not passed in with
`--spa`, but I think those are way too many keystrokes for a feature
someone would already expect

I manually tested this and it works great for every case